change key prefixes for SCT, seperate shieldedpool and sct methods
