+++ component/src/sct/state_key.rs
@@ -1,4 +1,3 @@
-use penumbra_crypto::{note, Nullifier};
 use penumbra_tct::{
     builder::{block, epoch},
     Root,
@@ -6,37 +5,29 @@ use penumbra_tct::{
 use std::string::String;
 
 pub fn anchor_by_height(height: u64) -> String {
-    format!("shielded_pool/anchor/{height}")
+    format!("sct/anchor/{height}")
 }
 
 pub fn anchor_lookup(anchor: Root) -> String {
-    format!("shielded_pool/valid_anchors/{anchor}")
+    format!("sct/valid_anchors/{anchor}")
 }
 
 pub fn stub_state_commitment_tree() -> &'static str {
-    "shielded_pool/stub/state_commitment_tree"
+    "sct/stub/state_commitment_tree"
 }
 
 pub fn block_anchor_by_height(height: u64) -> String {
-    format!("shielded_pool/block_anchor/{height}")
+    format!("sct/block_anchor/{height}")
 }
 
 pub fn block_anchor_lookup(anchor: block::Root) -> String {
-    format!("shielded_pool/valid_block_anchors/{anchor}")
-}
-
-pub fn note_source(note_commitment: &note::Commitment) -> String {
-    format!("shielded_pool/note_source/{note_commitment}")
+    format!("sct/valid_block_anchors/{anchor}")
 }
 
 pub fn epoch_anchor_lookup(anchor: epoch::Root) -> String {
-    format!("shielded_pool/valid_epoch_anchors/{anchor}")
-}
-
-pub fn spent_nullifier_lookup(nullifier: &Nullifier) -> String {
-    format!("shielded_pool/spent_nullifiers/{nullifier}")
+    format!("sct/valid_epoch_anchors/{anchor}")
 }
 
 pub fn epoch_anchor_by_index(index: u64) -> String {
-    format!("shielded_pool/epoch_anchor/{index}")
+    format!("sct/epoch_anchor/{index}")
 }
