+++ component/src/shielded_pool/state_key.rs
@@ -17,14 +17,6 @@ pub fn note_source(note_commitment: &note::Commitment) -> String {
     format!("shielded_pool/note_source/{note_commitment}")
 }
 
-pub fn compact_block(height: u64) -> String {
-    format!("shielded_pool/compact_block/{height}")
-}
-
-pub fn epoch_anchor_by_index(index: u64) -> String {
-    format!("shielded_pool/epoch_anchor/{index}")
-}
-
 pub fn spent_nullifier_lookup(nullifier: &Nullifier) -> String {
     format!("shielded_pool/spent_nullifiers/{nullifier}")
 }
